feat: Refactored and restructured profile view form
