+++ src/locales/en.json
@@ -1679,17 +1679,30 @@
     }
   },
   "profile": {
-    "title": "Agent persona",
-    "sub_title": "Customize the agent’s basic information and behavior.",
+    "description": "Customize the agent’s basic information and behavior.",
     "save_btn": "Save Changes",
     "invalid_field": "Required field",
     "invalid_get_data": "Error trying to get the data",
     "fields": {
-      "name": "Name",
-      "occupation": "Occupation",
-      "personality": "Personality",
-      "goal": "Goal",
-      "instruction": "Instruction",
+      "name": {
+        "title": "How does the agent introduce itself?",
+        "description": "This is the name displayed during interactions with users",
+        "placeholder": "Tainá"
+      },
+      "occupation": {
+        "title": "What’s the agent’s main role?",
+        "placeholder": "Customer Support Assistant",
+        "description": "Define the agent’s main role so users understand how it helps them"
+      },
+      "goal": {
+        "title": "What’s the agent’s main objective?",
+        "placeholder": "Answer customer questions",
+        "description": "This helps define what the agent aims to achieve during interactions"
+      },
+      "personality": {
+        "title": "What’s the agent’s tone of voice?",
+        "description": "Choose how the agent communicates with users"
+      },
       "personalities": {
         "friendly": "Friendly",
         "extrovert": "Extrovert",
@@ -1703,16 +1716,14 @@
         "intellectual": "Intellectual"
       }
     },
-    "placeholders": {
-      "instruction": "You're funny but you don't make jokes",
-      "name": "Tainá",
-      "occupation": "Llama expert", 
-      "goal": "To answer questions and give information about llamas"
-    },
     "instructions": {
       "title": "General instructions",
       "sub_title": "How you want the agent to behave.",
       "add_instruction_btn": "Add instruction",
+      "instruction": {
+        "title": "Instruction",
+        "placeholder": "You're funny but you don't make jokes"
+      },
       "modals": {
         "title": "Remove instruction",
         "description": "Are you sure you want to remove the instruction?",
