+++ src/locales/pt_br.json
@@ -1679,17 +1679,30 @@
     }
   },
   "profile": {
-    "title": "Persona do agente",
-    "sub_title": "Perfil do seu agente inteligente. Aqui é onde você irá inserir suas características e comportamentos.",
+    "description": "Perfil do seu agente inteligente. Aqui é onde você irá inserir suas características e comportamentos.",
     "save_btn": "Salvar Alterações",
     "invalid_field": "Campo obrigatório",
     "invalid_get_data": "Erro ao tentar buscar os dados",
     "fields": {
-      "name": "Nome",
-      "occupation": "Ocupação",
-      "personality": "Personalidade",
-      "goal": "Objetivo",
-      "instruction": "Instrução",
+      "name": {
+        "title": "Como o agente se apresenta?",
+        "description": "Este é o nome exibido durante as interações com os usuários",
+        "placeholder": "Tainá"
+      },
+      "occupation": {
+        "title": "Qual é a função principal do agente?",
+        "placeholder": "Assistente de Suporte ao Cliente",
+        "description": "Defina a função principal do agente para que os usuários entendam como isso os ajuda"
+      },
+      "goal": {
+        "title": "Qual é o objetivo principal do agente?",
+        "placeholder": "Responder perguntas dos clientes",
+        "description": "Isso ajuda a definir o que o agente pretende alcançar durante as interações"
+      },
+      "personality": {
+        "title": "Qual é o tom de voz do agente?",
+        "description": "Escolha como o agente se comunica com os usuários"
+      },
       "personalities": {
         "friendly": "Amigável",
         "extrovert": "Extrovertido",
@@ -1703,16 +1716,14 @@
         "intellectual": "Intelectual"
       }
     },
-    "placeholders": {
-      "instruction": "Você é divertido mas não faz piadas",
-      "name": "Tainá",
-      "occupation": "Especialista em lhamas",
-      "goal": "Tirar dúvidas e dar informações sobre lhamas"
-    },
     "instructions": {
       "title": "Instruções gerais",
       "sub_title": "Como você quer que o agente se comporte.",
       "add_instruction_btn": "Adicionar instrução",
+      "instruction": {
+        "title": "Instrução",
+        "placeholder": "Você é engraçado, mas não faz piadas"
+      },
       "modals": {
         "title": "Remover instrução",
         "description": "Tem certeza que deseja remover a instrução?",
